fix: actually remove all indirect usages of default unificator in :theory
