+++ theory/src/commonMain/kotlin/it/unibo/tuprolog/collections/rete/custom/Utils.kt
@@ -1,28 +1,13 @@
 package it.unibo.tuprolog.collections.rete.custom
 
-import it.unibo.tuprolog.collections.rete.custom.clause.IndexedClause
 import it.unibo.tuprolog.collections.rete.custom.clause.SituatedIndexedClause
 import it.unibo.tuprolog.collections.rete.takeFirstAfterSkipping
 import it.unibo.tuprolog.core.Clause
 import it.unibo.tuprolog.core.Struct
 import it.unibo.tuprolog.core.Term
-import it.unibo.tuprolog.unify.Unificator.Companion.matches
 import it.unibo.tuprolog.utils.mergeSequences
 
 internal object Utils {
-
-    fun <T : IndexedClause> removeAllLazily(source: MutableList<T>, clause: Clause): Sequence<T> =
-        sequence {
-            val iter = source.iterator()
-            while (iter.hasNext()) {
-                val it = iter.next()
-                if (it.innerClause matches clause) {
-                    it.invalidateAllCaches()
-                    iter.remove()
-                    yield(it)
-                }
-            }
-        }
 
     /**Calculate the arity of the first argument of any [Struct], at the given nesting level.
      * No checks are performed upon the validity of the Struct this extension method is called upon. */
