+++ theory/src/commonMain/kotlin/it/unibo/tuprolog/collections/rete/custom/nodes/ArityNode.kt
@@ -1,8 +1,7 @@
 package it.unibo.tuprolog.collections.rete.custom.nodes
 
+import it.unibo.tuprolog.collections.rete.custom.AbstractReteNode
 import it.unibo.tuprolog.collections.rete.custom.ReteNode
 import it.unibo.tuprolog.unify.Unificator
 
-internal abstract class ArityNode(
-    override val unificator: Unificator
-) : ReteNode
+internal abstract class ArityNode(unificator: Unificator) : ReteNode, AbstractReteNode(unificator)
