+++ theory/src/commonMain/kotlin/it/unibo/tuprolog/collections/rete/custom/nodes/ZeroArityReteNode.kt
@@ -1,12 +1,10 @@
 package it.unibo.tuprolog.collections.rete.custom.nodes
 
 import it.unibo.tuprolog.collections.rete.custom.Retractable
-import it.unibo.tuprolog.collections.rete.custom.Utils
 import it.unibo.tuprolog.collections.rete.custom.clause.IndexedClause
 import it.unibo.tuprolog.collections.rete.custom.clause.SituatedIndexedClause
 import it.unibo.tuprolog.core.Clause
 import it.unibo.tuprolog.unify.Unificator
-import it.unibo.tuprolog.unify.Unificator.Companion.matches
 import it.unibo.tuprolog.utils.addFirst
 import it.unibo.tuprolog.utils.buffered
 import it.unibo.tuprolog.utils.dequeOf
@@ -20,7 +18,7 @@ internal class ZeroArityReteNode(
         dequeOf()
 
     override fun retractFirst(clause: Clause): Sequence<Clause> =
-        Utils.removeAllLazily(atoms, clause).map { it.innerClause }.take(1).buffered()
+        removeAllLazily(atoms, clause).map { it.innerClause }.take(1).buffered()
 
     override val size: Int
         get() = atoms.size
@@ -29,7 +27,7 @@ internal class ZeroArityReteNode(
         get() = atoms.isEmpty()
 
     override fun get(clause: Clause): Sequence<Clause> =
-        atoms.asSequence().filter { it.innerClause matches clause }.map { it.innerClause }
+        atoms.asSequence().filter { unificator.match(it.innerClause, clause) }.map { it.innerClause }
 
     override fun assertA(clause: IndexedClause) {
         if (ordered) {
@@ -54,7 +52,7 @@ internal class ZeroArityReteNode(
     }
 
     override fun retractAll(clause: Clause): Sequence<Clause> =
-        Utils.removeAllLazily(atoms, clause).map { it.innerClause }.buffered()
+        removeAllLazily(atoms, clause).map { it.innerClause }.buffered()
 
     override fun getCache(): Sequence<SituatedIndexedClause> {
         return atoms.asSequence()
